mm, compaction: Skip all non-migratable pages during scan
Pages pinned in memory through extra refcounts can not be migrated. 
Currently as isolate_migratepages_block() scans pages for compaction, it
skips any pinned anonymous pages.  All non-migratable pages should be
skipped and not just the anonymous pinned pages.

This patch adds a check for extra refcounts on a page to determine if
the page can be migrated.  This was seen as a real issue on a customer
workload where a large number of pages were pinned by vfio on the host
and any attempts to allocate hugepages resulted in significant amount
of cpu time spent in either direct compaction or in kcompactd scanning
vfio pinned pages over and over again that can not be migrated.

Link: https://lkml.kernel.org/r/20230512190000.103737-1-khalid.aziz@oracle.com
Signed-off-by: Khalid Aziz <khalid.aziz@oracle.com>
Suggested-by: Steve Sistare <steven.sistare@oracle.com>
Cc: Matthew Wilcox <willy@infradead.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>